+++ src/common/types/hyperloglog.cpp
@@ -17,14 +17,14 @@ HyperLogLog::~HyperLogLog() {
 	hll_destroy((robj *)hll);
 }
 
-void HyperLogLog::Add(uint8_t *element, size_t size) {
+void HyperLogLog::Add(uint8_t *element, uint64_t size) {
 	if (hll_add((robj *)hll, element, size) == C_ERR) {
 		throw Exception("Could not add to HLL?");
 	}
 }
 
-size_t HyperLogLog::Count() {
-	size_t result;
+uint64_t HyperLogLog::Count() {
+	size_t result; // exception from size_t ban
 	if (hll_count((robj *)hll, &result) != C_OK) {
 		throw Exception("Could not count HLL?");
 	}
@@ -42,10 +42,10 @@ unique_ptr<HyperLogLog> HyperLogLog::Merge(HyperLogLog &other) {
 	return unique_ptr<HyperLogLog>(new HyperLogLog((void *)new_hll));
 }
 
-unique_ptr<HyperLogLog> HyperLogLog::Merge(HyperLogLog logs[], size_t count) {
+unique_ptr<HyperLogLog> HyperLogLog::Merge(HyperLogLog logs[], uint64_t count) {
 	auto hlls_uptr = unique_ptr<robj *[]> { new robj *[count] };
 	auto hlls = hlls_uptr.get();
-	for (size_t i = 0; i < count; i++) {
+	for (uint64_t i = 0; i < count; i++) {
 		hlls[i] = (robj *)logs[i].hll;
 	}
 	auto new_hll = hll_merge(hlls, count);
