+++ src/common/types/string_heap.cpp
@@ -10,7 +10,7 @@ using namespace std;
 StringHeap::StringHeap() : tail(nullptr) {
 }
 
-const char *StringHeap::AddString(const char *data, size_t len) {
+const char *StringHeap::AddString(const char *data, uint64_t len) {
 #ifdef DEBUG
 	if (!Value::IsUTF8String(data)) {
 		throw Exception("String value is not valid UTF8");
@@ -18,7 +18,7 @@ const char *StringHeap::AddString(const char *data, size_t len) {
 #endif
 	if (!chunk || chunk->current_position + len >= chunk->maximum_size) {
 		// have to make a new entry
-		auto new_chunk = make_unique<StringChunk>(std::max(len + 1, (size_t)MINIMUM_HEAP_SIZE));
+		auto new_chunk = make_unique<StringChunk>(std::max(len + 1, (uint64_t)MINIMUM_HEAP_SIZE));
 		new_chunk->prev = move(chunk);
 		chunk = move(new_chunk);
 		if (!tail) {
