+++ src/include/function/scalar_function/regexp_matches.hpp
@@ -14,7 +14,7 @@
 
 namespace duckdb {
 
-void regexp_matches_function(ExpressionExecutor &exec, Vector inputs[], size_t input_count,
+void regexp_matches_function(ExpressionExecutor &exec, Vector inputs[], uint64_t input_count,
                              BoundFunctionExpression &expr, Vector &result);
 bool regexp_matches_matches_arguments(vector<SQLType> &arguments);
 SQLType regexp_matches_get_return_type(vector<SQLType> &arguments);
